requirements.txt
easydict
matplotlib==3.0.0
numpy==1.16.1
opencv-python==3.4.4.19
Pillow==6.2.0
scipy==1.1.0
tensorflow
tensorboard==1.9.0
tensorboardX==1.6
torch==1.2.0
torchvision==0.3.0
tqdm==4.25.0